refactor(maintenance): split large source files
