+++ lib/presentation/widgets/image_editor/image_editor_screen.dart
@@ -35,107 +35,10 @@ import 'canvas/layer_painter.dart';
 import 'export/image_exporter_new.dart';
 import '../../widgets/common/themed_divider.dart';
 
-enum ImageEditorMode { edit, inpaint }
-
-class ImageEditorFocusedInpaintCostConfig {
-  const ImageEditorFocusedInpaintCostConfig({
-    required this.model,
-    required this.steps,
-    required this.batchCount,
-    required this.batchSize,
-    required this.smea,
-    required this.smeaDyn,
-    required this.subscriptionTier,
-    this.extraPerSampleCost = 0,
-  });
-
-  final String model;
-  final int steps;
-  final int batchCount;
-  final int batchSize;
-  final bool smea;
-  final bool smeaDyn;
-  final int subscriptionTier;
-  final int extraPerSampleCost;
-
-  int estimate({required int width, required int height}) {
-    return AnlasCalculator.calculateRequestCost(
-      width: width,
-      height: height,
-      steps: steps,
-      batchCount: batchCount,
-      batchSize: batchSize,
-      smea: smea,
-      smeaDyn: smeaDyn,
-      model: model,
-      subscriptionTier: subscriptionTier,
-      strength: 1.0,
-      extraPerSampleCost: extraPerSampleCost,
-    );
-  }
-}
-
-class _FocusedInpaintCostEstimate {
-  const _FocusedInpaintCostEstimate({
-    required this.width,
-    required this.height,
-    required this.cost,
-  });
-
-  final int width;
-  final int height;
-  final int cost;
-}
-
-/// 图像编辑器返回结果
-class ImageEditorResult {
-  /// 修改后的图像（涂鸦合并）
-  final Uint8List? modifiedImage;
-
-  /// Inpainting蒙版图像
-  final Uint8List? maskImage;
-
-  /// 是否有图像修改
-  final bool hasImageChanges;
-
-  /// 是否有蒙版修改
-  final bool hasMaskChanges;
-
-  /// Focused Inpaint 选区范围
-  final Rect? focusAreaRect;
-
-  /// Focused Inpaint 上下文带宽
-  final double minimumContextMegaPixels;
-
-  /// 是否启用 Focused Inpaint
-  final bool focusedInpaintEnabled;
-
-  /// Outpaint 扩展后的源图像
-  final Uint8List? outpaintSourceImage;
-
-  /// Outpaint 扩展后的源图像宽度
-  final int? outpaintSourceWidth;
-
-  /// Outpaint 扩展后的源图像高度
-  final int? outpaintSourceHeight;
-
-  /// 是否有 Outpaint 源图像修改
-  final bool hasOutpaintChanges;
-
-  const ImageEditorResult({
-    this.modifiedImage,
-    this.maskImage,
-    this.hasImageChanges = false,
-    this.hasMaskChanges = false,
-    this.focusAreaRect,
-    this.minimumContextMegaPixels = 88.0,
-    this.focusedInpaintEnabled = false,
-    this.outpaintSourceImage,
-    this.outpaintSourceWidth,
-    this.outpaintSourceHeight,
-    this.hasOutpaintChanges = false,
-  });
-}
+part 'image_editor_screen_effects.dart';
+part 'image_editor_screen_focused.dart';
+part 'image_editor_screen_layout.dart';
+part 'image_editor_screen_types.dart';
 
 /// 图像编辑器主界面
 class ImageEditorScreen extends StatefulWidget {
@@ -627,1271 +530,8 @@ class _ImageEditorScreenState extends State<ImageEditorScreen> {
     );
   }
 
-  /// 桌面端布局
-  Widget _buildDesktopLayout() {
-    return Scaffold(
-      body: Column(
-        children: [
-          // 顶部菜单栏
-          _buildDesktopMenuBar(),
+  void _updateLayoutState(VoidCallback update) => setState(update);
 
-          // 主体区域
-          Expanded(
-            child: Row(
-              children: [
-                // 左侧工具栏
-                DesktopToolbar(
-                  state: _state,
-                  onClear: _isInpaintMode ? _resetInpaintMask : null,
-                  onFillMask: _isInpaintMode
-                      ? _handleFillClosedMaskRegions
-                      : null,
-                  canFillMask: _isInpaintMode ? _hasMaskContent : null,
-                  allowedToolIds: _isInpaintMode ? _inpaintToolIds : null,
-                ),
-
-                // 中间画布区域
-                Expanded(
-                  child: Column(
-                    children: [
-                      Expanded(child: _buildCanvasArea()),
-                      // 底部状态栏
-                      _buildStatusBar(),
-                    ],
-                  ),
-                ),
-
-                // 右侧面板
-                if (_showLayerPanel)
-                  SizedBox(
-                    width: 280,
-                    child: Column(
-                      children: [
-                        // 图层面板
-                        Expanded(flex: 2, child: LayerPanel(state: _state)),
-                        const ThemedDivider(height: 1),
-                        // 工具设置面板
-                        Expanded(flex: 2, child: _buildToolSettingsPanel()),
-                        const ThemedDivider(height: 1),
-                        // 颜色面板
-                        if (!_isInpaintMode) ColorPanel(state: _state),
-                      ],
-                    ),
-                  ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// 移动端布局
-  Widget _buildMobileLayout() {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(_editorTitle()),
-        actions: [
-          // 图层按钮
-          IconButton(
-            icon: const Icon(Icons.layers),
-            onPressed: _showMobileLayerSheet,
-            tooltip: context.l10n.editor_layers,
-          ),
-          // 加载蒙版按钮
-          if (_isInpaintMode)
-            IconButton(
-              icon: const Icon(Icons.upload_file),
-              onPressed: _loadMask,
-              tooltip: context.l10n.editor_loadMask,
-            ),
-          if (_isInpaintMode)
-            IconButton(
-              icon: const Icon(Icons.open_in_full),
-              onPressed: _showShiftEdgesDialog,
-              tooltip: 'Shift Edges',
-            ),
-          if (!_isInpaintMode)
-            IconButton(
-              icon: const Icon(Icons.tune_rounded),
-              onPressed: _showEffectsDialog,
-              tooltip: 'Effects',
-            ),
-          // 导出按钮
-          IconButton(
-            icon: const Icon(Icons.check),
-            onPressed: _canExportAndClose ? _exportAndClose : null,
-            tooltip: context.l10n.editor_done,
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          // 画布区域
-          Expanded(child: _buildCanvasArea()),
-
-          // 工具设置（可折叠）
-          _buildMobileToolSettings(),
-
-          // 底部工具栏
-          MobileToolbar(
-            state: _state,
-            onClear: _isInpaintMode ? _resetInpaintMask : null,
-            onFillMask: _isInpaintMode ? _handleFillClosedMaskRegions : null,
-            canFillMask: _isInpaintMode ? _hasMaskContent : null,
-            onLayersPressed: _showMobileLayerSheet,
-            allowedToolIds: _isInpaintMode ? _inpaintToolIds : null,
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// 桌面端菜单栏
-  Widget _buildDesktopMenuBar() {
-    final theme = Theme.of(context);
-
-    return Container(
-      height: 40,
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surface,
-        border: Border(
-          bottom: BorderSide(color: theme.dividerColor.withValues(alpha: 0.3)),
-        ),
-      ),
-      child: Row(
-        children: [
-          // 返回按钮
-          IconButton(
-            icon: const Icon(Icons.arrow_back, size: 20),
-            onPressed: () => _confirmExit(),
-            tooltip: context.l10n.editor_back,
-          ),
-
-          Text(_editorTitle(), style: theme.textTheme.titleSmall),
-
-          const Spacer(),
-
-          if (!_isInpaintMode)
-            TextButton.icon(
-              icon: const Icon(Icons.tune_rounded, size: 18),
-              label: const Text('Effects'),
-              onPressed: _showEffectsDialog,
-            ),
-
-          // 画布尺寸按钮（使用细粒度监听）
-          TextButton.icon(
-            icon: const Icon(Icons.aspect_ratio, size: 18),
-            label: ValueListenableBuilder<Size>(
-              valueListenable: _state.canvasSizeNotifier,
-              builder: (context, size, _) =>
-                  Text('${size.width.toInt()} x ${size.height.toInt()}'),
-            ),
-            onPressed: _changeCanvasSize,
-          ),
-
-          // 加载蒙版按钮
-          if (_isInpaintMode)
-            IconButton(
-              icon: const Icon(Icons.upload_file, size: 20),
-              onPressed: _loadMask,
-              tooltip: context.l10n.editor_loadMask,
-            ),
-
-          if (_isInpaintMode)
-            TextButton.icon(
-              icon: const Icon(Icons.open_in_full, size: 18),
-              label: const Text('Shift Edges'),
-              onPressed: _showShiftEdgesDialog,
-            ),
-
-          const ThemedDivider(
-            height: 1,
-            vertical: true,
-            indent: 8,
-            endIndent: 8,
-          ),
-
-          // 切换面板
-          IconButton(
-            icon: Icon(
-              _showLayerPanel
-                  ? Icons.view_sidebar
-                  : Icons.view_sidebar_outlined,
-              size: 20,
-            ),
-            onPressed: () {
-              setState(() {
-                _showLayerPanel = !_showLayerPanel;
-              });
-            },
-            tooltip: context.l10n.editor_togglePanels,
-          ),
-
-          // 快捷键帮助
-          IconButton(
-            icon: const Icon(Icons.keyboard, size: 20),
-            onPressed: _showShortcutHelp,
-            tooltip: context.l10n.editor_shortcutHelpTitle,
-          ),
-
-          const ThemedDivider(
-            height: 1,
-            vertical: true,
-            indent: 8,
-            endIndent: 8,
-          ),
-
-          // 导出按钮
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 8),
-            child: FilledButton.icon(
-              icon: const Icon(Icons.check, size: 18),
-              label: Text(context.l10n.editor_done),
-              onPressed: _canExportAndClose ? _exportAndClose : null,
-              style: FilledButton.styleFrom(
-                padding: const EdgeInsets.symmetric(horizontal: 16),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// 状态栏
-  /// 使用 Listenable.merge 实现细粒度监听
-  Widget _buildStatusBar() {
-    final theme = Theme.of(context);
-
-    return ListenableBuilder(
-      listenable: Listenable.merge([
-        _state.canvasController, // 缩放、旋转、镜像
-        _state.canvasSizeNotifier, // 画布尺寸
-        _state.layerManager, // 图层数量
-        _state.selectionManager, // 选区状态
-      ]),
-      builder: (context, _) {
-        return Container(
-          height: 24,
-          padding: const EdgeInsets.symmetric(horizontal: 12),
-          decoration: BoxDecoration(
-            color: theme.colorScheme.surfaceContainerHighest,
-            border: Border(
-              top: BorderSide(color: theme.dividerColor.withValues(alpha: 0.3)),
-            ),
-          ),
-          child: Row(
-            children: [
-              Text(
-                context.l10n.editor_statusZoom(
-                  (_state.canvasController.scale * 100).round(),
-                ),
-                style: theme.textTheme.bodySmall,
-              ),
-              const SizedBox(width: 16),
-              Text(
-                context.l10n.editor_statusCanvas(
-                  _state.canvasSize.width.toInt(),
-                  _state.canvasSize.height.toInt(),
-                ),
-                style: theme.textTheme.bodySmall,
-              ),
-              const SizedBox(width: 16),
-              Text(
-                context.l10n.editor_statusLayers(
-                  _state.layerManager.layerCount,
-                ),
-                style: theme.textTheme.bodySmall,
-              ),
-              if (_state.selectionPath != null) ...[
-                const SizedBox(width: 16),
-                Text(
-                  context.l10n.editor_statusHasSelection,
-                  style: theme.textTheme.bodySmall?.copyWith(
-                    color: theme.colorScheme.primary,
-                  ),
-                ),
-              ],
-              // 旋转角度显示
-              if (_state.canvasController.rotation != 0) ...[
-                const SizedBox(width: 16),
-                Text(
-                  context.l10n.editor_statusRotation(
-                    (_state.canvasController.rotation * 180 / 3.14159265359)
-                        .round(),
-                  ),
-                  style: theme.textTheme.bodySmall?.copyWith(
-                    color: theme.colorScheme.secondary,
-                  ),
-                ),
-              ],
-              // 镜像状态显示
-              if (_state.canvasController.isMirroredHorizontally) ...[
-                const SizedBox(width: 16),
-                Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Icon(
-                      Icons.flip,
-                      size: 14,
-                      color: theme.colorScheme.secondary,
-                    ),
-                    const SizedBox(width: 4),
-                    Text(
-                      context.l10n.editor_statusMirrored,
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        color: theme.colorScheme.secondary,
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  /// 工具设置面板
-  /// 使用 toolChangeNotifier 实现细粒度监听，仅在工具切换时重建
-  Widget _buildToolSettingsPanel() {
-    return ValueListenableBuilder<EditorTool?>(
-      valueListenable: _state.toolChangeNotifier,
-      builder: (context, tool, _) {
-        if (tool == null) {
-          return Center(child: Text(context.l10n.image_editor_select_tool));
-        }
-        return SingleChildScrollView(
-          child: tool.buildSettingsPanel(context, _state),
-        );
-      },
-    );
-  }
-
-  /// 移动端工具设置
-  /// 使用 toolChangeNotifier 实现细粒度监听
-  Widget _buildMobileToolSettings() {
-    return ValueListenableBuilder<EditorTool?>(
-      valueListenable: _state.toolChangeNotifier,
-      builder: (context, tool, _) {
-        if (tool == null) return const SizedBox.shrink();
-
-        return Container(
-          constraints: const BoxConstraints(maxHeight: 150),
-          child: SingleChildScrollView(
-            child: tool.buildSettingsPanel(context, _state),
-          ),
-        );
-      },
-    );
-  }
-
-  /// 显示移动端图层面板
-  void _showMobileLayerSheet() {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) => DraggableScrollableSheet(
-        initialChildSize: 0.5,
-        minChildSize: 0.3,
-        maxChildSize: 0.9,
-        expand: false,
-        builder: (context, scrollController) {
-          return LayerPanel(state: _state);
-        },
-      ),
-    );
-  }
-
-  /// 显示快捷键帮助
-  void _showShortcutHelp() {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: Row(
-          children: [
-            const Icon(Icons.keyboard),
-            const SizedBox(width: 8),
-            Text(context.l10n.editor_shortcutHelpTitle),
-          ],
-        ),
-        content: ConstrainedBox(
-          constraints: const BoxConstraints(maxHeight: 500, maxWidth: 350),
-          child: SingleChildScrollView(
-            primary: true,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                _buildShortcutSection(context.l10n.editor_shortcutPaintTools, [
-                  ('B', context.l10n.editor_toolBrush),
-                  ('E', context.l10n.editor_toolEraser),
-                  ('P', context.l10n.editor_toolColorPicker),
-                  ('Alt', context.l10n.editor_shortcutTemporaryColorPicker),
-                ]),
-                _buildShortcutSection(
-                  context.l10n.editor_shortcutSelectionTools,
-                  [
-                    ('M', context.l10n.editor_shortcutRectSelection),
-                    ('U', context.l10n.editor_shortcutEllipseSelection),
-                    ('L', context.l10n.editor_shortcutLassoSelection),
-                  ],
-                ),
-                _buildShortcutSection(context.l10n.editor_shortcutCanvasView, [
-                  ('1', context.l10n.editor_shortcut100Zoom),
-                  ('2', context.l10n.editor_shortcutFitHeight),
-                  ('3', context.l10n.editor_shortcutFitWidth),
-                  ('4', context.l10n.editor_shortcutRotateLeft15),
-                  ('5', context.l10n.editor_shortcutResetRotation),
-                  ('6', context.l10n.editor_shortcutRotateRight15),
-                  ('F', context.l10n.editor_shortcutFlipHorizontal),
-                  ('R', context.l10n.editor_resetView),
-                  (context.l10n.editor_shortcutWheel, context.l10n.editor_zoom),
-                  ('Ctrl+0', context.l10n.editor_shortcut100Zoom),
-                  ('Ctrl++', context.l10n.editor_zoomIn),
-                  ('Ctrl+-', context.l10n.editor_zoomOut),
-                ]),
-                _buildShortcutSection(context.l10n.editor_shortcutBrushAdjust, [
-                  ('[', context.l10n.editor_shortcutBrushSmaller),
-                  (']', context.l10n.editor_shortcutBrushLarger),
-                  ('I', context.l10n.editor_shortcutOpacityLower),
-                  ('O', context.l10n.editor_shortcutOpacityHigher),
-                  ('Shift + Drag', context.l10n.editor_shortcutDragBrushSize),
-                ]),
-                _buildShortcutSection(context.l10n.editor_shortcutColors, [
-                  ('X', context.l10n.editor_shortcutSwapColors),
-                ]),
-                _buildShortcutSection(
-                  context.l10n.editor_shortcutCanvasActions,
-                  [
-                    ('Space + Drag', context.l10n.editor_shortcutPanCanvas),
-                    ('Middle Drag', context.l10n.editor_shortcutPanCanvas),
-                  ],
-                ),
-                _buildShortcutSection(
-                  context.l10n.editor_shortcutHistoryActions,
-                  [
-                    ('Ctrl+Z', context.l10n.editor_undo),
-                    ('Ctrl+Shift+Z', context.l10n.editor_redo),
-                    ('Ctrl+Y', context.l10n.editor_redo),
-                  ],
-                ),
-                _buildShortcutSection(
-                  context.l10n.editor_shortcutSelectionActions,
-                  [
-                    (
-                      'Delete',
-                      context.l10n.editor_shortcutClearSelectionContent,
-                    ),
-                    (
-                      'Backspace',
-                      context.l10n.editor_shortcutClearSelectionContent,
-                    ),
-                    ('Esc', context.l10n.editor_shortcutCancelCurrentAction),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: Text(context.l10n.common_close),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _showEffectsDialog() async {
-    final layer = _state.layerManager.activeLayer;
-    if (layer == null || layer.locked || !layer.hasContent) {
-      AppToast.warning(
-        context,
-        context.l10n.editor_selectUnlockedLayerWithContent,
-      );
-      return;
-    }
-
-    final sourceBytes = await _readLayerPng(layer);
-    if (!mounted) return;
-    if (sourceBytes == null) {
-      AppToast.error(context, context.l10n.editor_readCurrentLayerFailed);
-      return;
-    }
-
-    var effectType = EditorEffectType.brightness;
-    var intensity = 0.25;
-    var previewBytes = sourceBytes;
-    var previewLoading = false;
-    var previewError = '';
-    var previewVersion = 0;
-    var previewInitialized = false;
-    var dialogOpen = true;
-    Timer? previewDebounce;
-
-    Future<void> refreshPreview(StateSetter setDialogState) async {
-      previewDebounce?.cancel();
-      final version = ++previewVersion;
-      setDialogState(() {
-        previewLoading = true;
-        previewError = '';
-      });
-
-      previewDebounce = Timer(const Duration(milliseconds: 180), () async {
-        try {
-          final cropRect = _selectionCropRect();
-          final job = EditorEffectJob(
-            imageBytes: sourceBytes,
-            effectType: effectType,
-            intensity: intensity,
-            maxPreviewDimension: 768,
-            cropRect: cropRect,
-          );
-          final resultMessage = await compute(
-            runEditorEffectJobMessage,
-            job.toMessage(),
-            debugLabel: 'image_editor_effect_preview',
-          );
-          final result = EditorEffectResult.fromMessage(resultMessage);
-          if (!dialogOpen || !mounted || version != previewVersion) {
-            return;
-          }
-          setDialogState(() {
-            previewBytes = result.bytes;
-            previewLoading = false;
-          });
-        } catch (e) {
-          if (!dialogOpen || !mounted || version != previewVersion) {
-            return;
-          }
-          setDialogState(() {
-            previewLoading = false;
-            previewError = e.toString();
-          });
-        }
-      });
-    }
-
-    final confirmed = await showDialog<bool>(
-      context: context,
-      builder: (context) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            if (!previewInitialized) {
-              previewInitialized = true;
-              WidgetsBinding.instance.addPostFrameCallback((_) {
-                if (dialogOpen && mounted) {
-                  unawaited(refreshPreview(setState));
-                }
-              });
-            }
-            final media = MediaQuery.of(context);
-            final horizontalInset = media.size.width < 820 ? 12.0 : 32.0;
-            final dialogWidth = (media.size.width - horizontalInset * 2)
-                .clamp(360.0, 1120.0)
-                .toDouble();
-            final previewHeight = (media.size.height * 0.48)
-                .clamp(320.0, 520.0)
-                .toDouble();
-
-            void selectEffect(EditorEffectType value) {
-              setState(() {
-                effectType = value;
-                intensity = _defaultEffectIntensity(value);
-              });
-              unawaited(refreshPreview(setState));
-            }
-
-            return Dialog(
-              insetPadding: EdgeInsets.symmetric(
-                horizontal: horizontalInset,
-                vertical: 20,
-              ),
-              child: ConstrainedBox(
-                constraints: BoxConstraints(
-                  maxWidth: dialogWidth,
-                  maxHeight: media.size.height * 0.9,
-                ),
-                child: Padding(
-                  padding: const EdgeInsets.all(24),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Row(
-                        children: [
-                          Expanded(
-                            child: Text(
-                              context.l10n.editor_localEffects,
-                              style: Theme.of(context).textTheme.headlineSmall,
-                            ),
-                          ),
-                          IconButton(
-                            tooltip: context.l10n.common_close,
-                            onPressed: () => Navigator.pop(context, false),
-                            icon: const Icon(Icons.close),
-                          ),
-                        ],
-                      ),
-                      const SizedBox(height: 16),
-                      Flexible(
-                        child: SingleChildScrollView(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              _buildEffectSection(
-                                title: context.l10n.editor_basicAdjustments,
-                                effects: const [
-                                  EditorEffectType.brightness,
-                                  EditorEffectType.contrast,
-                                  EditorEffectType.saturation,
-                                  EditorEffectType.temperature,
-                                  EditorEffectType.gamma,
-                                ],
-                                selectedEffect: effectType,
-                                onSelected: selectEffect,
-                              ),
-                              const SizedBox(height: 14),
-                              _buildEffectSection(
-                                title: context.l10n.editor_styleAndRepair,
-                                effects: const [
-                                  EditorEffectType.grayscale,
-                                  EditorEffectType.invert,
-                                  EditorEffectType.sepia,
-                                  EditorEffectType.denoise,
-                                  EditorEffectType.blur,
-                                  EditorEffectType.sharpen,
-                                ],
-                                selectedEffect: effectType,
-                                onSelected: selectEffect,
-                              ),
-                              const SizedBox(height: 14),
-                              _buildEffectSection(
-                                title: context.l10n.editor_transformCrop,
-                                description: context
-                                    .l10n
-                                    .editor_transformCropDescription,
-                                effects: const [
-                                  EditorEffectType.rotateLeft,
-                                  EditorEffectType.rotateRight,
-                                  EditorEffectType.flipHorizontal,
-                                  EditorEffectType.flipVertical,
-                                  EditorEffectType.cropToSelection,
-                                ],
-                                selectedEffect: effectType,
-                                onSelected: selectEffect,
-                                prominent: true,
-                              ),
-                              const SizedBox(height: 16),
-                              _buildEffectControl(
-                                effectType: effectType,
-                                intensity: intensity,
-                                onChanged: (value) {
-                                  setState(() => intensity = value);
-                                  unawaited(refreshPreview(setState));
-                                },
-                                onReset: () {
-                                  setState(
-                                    () => intensity = _defaultEffectIntensity(
-                                      effectType,
-                                    ),
-                                  );
-                                  unawaited(refreshPreview(setState));
-                                },
-                              ),
-                              const SizedBox(height: 16),
-                              _buildEffectPreviewComparison(
-                                previewHeight: previewHeight,
-                                sourceBytes: sourceBytes,
-                                previewBytes: previewBytes,
-                                previewLoading: previewLoading,
-                                previewError: previewError,
-                              ),
-                              const SizedBox(height: 12),
-                              Text(
-                                context.l10n.editor_effectPreviewHint,
-                                style: Theme.of(context).textTheme.bodyMedium,
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 18),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          TextButton(
-                            onPressed: () => Navigator.pop(context, false),
-                            child: Text(context.l10n.common_cancel),
-                          ),
-                          const SizedBox(width: 12),
-                          FilledButton.icon(
-                            onPressed: previewLoading || previewError.isNotEmpty
-                                ? null
-                                : () => Navigator.pop(context, true),
-                            icon: const Icon(Icons.check),
-                            label: Text(
-                              context.l10n.editor_applyToCurrentLayer,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            );
-          },
-        );
-      },
-    );
-    dialogOpen = false;
-    previewDebounce?.cancel();
-
-    if (confirmed == true) {
-      await _applyEffect(effectType, intensity);
-    }
-  }
-
-  Widget _buildEffectSection({
-    required String title,
-    required List<EditorEffectType> effects,
-    required EditorEffectType selectedEffect,
-    required ValueChanged<EditorEffectType> onSelected,
-    String? description,
-    bool prominent = false,
-  }) {
-    final theme = Theme.of(context);
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surfaceContainerHighest.withValues(alpha: 0.3),
-        borderRadius: BorderRadius.circular(16),
-        border: Border.all(color: theme.colorScheme.outlineVariant),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(14),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Text(
-                  title,
-                  style: theme.textTheme.titleSmall?.copyWith(
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                if (description != null) ...[
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      description,
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        color: theme.colorScheme.onSurfaceVariant,
-                      ),
-                    ),
-                  ),
-                ],
-              ],
-            ),
-            const SizedBox(height: 10),
-            Wrap(
-              spacing: 8,
-              runSpacing: 8,
-              children: [
-                for (final effect in effects)
-                  _buildEffectChip(
-                    effect: effect,
-                    selected: effect == selectedEffect,
-                    onSelected: onSelected,
-                    prominent: prominent,
-                  ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildEffectChip({
-    required EditorEffectType effect,
-    required bool selected,
-    required ValueChanged<EditorEffectType> onSelected,
-    required bool prominent,
-  }) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final foreground = selected
-        ? colorScheme.onSecondaryContainer
-        : colorScheme.onSurface;
-    return ChoiceChip(
-      selected: selected,
-      showCheckmark: false,
-      selectedColor: colorScheme.secondaryContainer,
-      backgroundColor: colorScheme.surfaceContainerHighest.withValues(
-        alpha: 0.55,
-      ),
-      side: BorderSide(
-        color: selected ? colorScheme.secondary : colorScheme.outlineVariant,
-      ),
-      padding: EdgeInsets.symmetric(
-        horizontal: prominent ? 14 : 10,
-        vertical: prominent ? 10 : 7,
-      ),
-      label: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(_effectIcon(effect), size: prominent ? 20 : 18),
-          const SizedBox(width: 6),
-          Text(
-            _effectLabel(effect),
-            style: theme.textTheme.labelLarge?.copyWith(color: foreground),
-          ),
-        ],
-      ),
-      onSelected: (_) => onSelected(effect),
-    );
-  }
-
-  Widget _buildEffectControl({
-    required EditorEffectType effectType,
-    required double intensity,
-    required ValueChanged<double> onChanged,
-    required VoidCallback onReset,
-  }) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    if (!_effectHasIntensity(effectType)) {
-      return DecoratedBox(
-        decoration: BoxDecoration(
-          color: colorScheme.surfaceContainerHighest.withValues(alpha: 0.35),
-          borderRadius: BorderRadius.circular(16),
-          border: Border.all(color: colorScheme.outlineVariant),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(14),
-          child: Row(
-            children: [
-              Icon(_effectIcon(effectType), color: colorScheme.primary),
-              const SizedBox(width: 10),
-              Expanded(
-                child: Text(
-                  context.l10n.editor_oneShotEffectHint(
-                    _effectLabel(effectType),
-                  ),
-                  style: theme.textTheme.bodyMedium,
-                ),
-              ),
-            ],
-          ),
-        ),
-      );
-    }
-
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: colorScheme.surfaceContainerHighest.withValues(alpha: 0.35),
-        borderRadius: BorderRadius.circular(16),
-        border: Border.all(color: colorScheme.outlineVariant),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(14, 12, 14, 8),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Icon(_effectIcon(effectType), color: colorScheme.primary),
-                const SizedBox(width: 10),
-                Expanded(
-                  child: Text(
-                    context.l10n.editor_effectIntensity(
-                      _effectLabel(effectType),
-                    ),
-                    style: theme.textTheme.titleSmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-                Text(
-                  intensity.toStringAsFixed(2),
-                  style: theme.textTheme.titleSmall,
-                ),
-                const SizedBox(width: 8),
-                TextButton(
-                  onPressed: onReset,
-                  child: Text(context.l10n.common_reset),
-                ),
-              ],
-            ),
-            Slider(
-              value: intensity,
-              min: _effectMin(effectType),
-              max: _effectMax(effectType),
-              divisions: 40,
-              onChanged: onChanged,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildEffectPreviewComparison({
-    required double previewHeight,
-    required Uint8List sourceBytes,
-    required Uint8List previewBytes,
-    required bool previewLoading,
-    required String previewError,
-  }) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        final stacked = constraints.maxWidth < 720;
-        if (stacked) {
-          return SizedBox(
-            height: previewHeight * 1.7,
-            child: Column(
-              children: [
-                Expanded(
-                  child: _buildEffectPreviewPane(
-                    title: context.l10n.editor_original,
-                    bytes: sourceBytes,
-                    loading: false,
-                    error: '',
-                  ),
-                ),
-                const SizedBox(height: 12),
-                Expanded(
-                  child: _buildEffectPreviewPane(
-                    title: context.l10n.editor_effectPreview,
-                    bytes: previewBytes,
-                    loading: previewLoading,
-                    error: previewError,
-                  ),
-                ),
-              ],
-            ),
-          );
-        }
-
-        return SizedBox(
-          height: previewHeight,
-          child: Row(
-            children: [
-              Expanded(
-                child: _buildEffectPreviewPane(
-                  title: context.l10n.editor_original,
-                  bytes: sourceBytes,
-                  loading: false,
-                  error: '',
-                ),
-              ),
-              const SizedBox(width: 14),
-              Expanded(
-                child: _buildEffectPreviewPane(
-                  title: context.l10n.editor_effectPreview,
-                  bytes: previewBytes,
-                  loading: previewLoading,
-                  error: previewError,
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildEffectPreviewPane({
-    required String title,
-    required Uint8List bytes,
-    required bool loading,
-    required String error,
-  }) {
-    final theme = Theme.of(context);
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surfaceContainerHighest.withValues(alpha: 0.5),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: theme.colorScheme.outlineVariant),
-      ),
-      child: Stack(
-        children: [
-          Positioned.fill(
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(8, 28, 8, 8),
-              child: error.isNotEmpty
-                  ? Center(
-                      child: Text(
-                        error,
-                        maxLines: 4,
-                        overflow: TextOverflow.ellipsis,
-                        textAlign: TextAlign.center,
-                        style: theme.textTheme.bodySmall?.copyWith(
-                          color: theme.colorScheme.error,
-                        ),
-                      ),
-                    )
-                  : Image.memory(
-                      bytes,
-                      fit: BoxFit.contain,
-                      filterQuality: FilterQuality.medium,
-                      gaplessPlayback: true,
-                    ),
-            ),
-          ),
-          Positioned(
-            left: 8,
-            top: 6,
-            child: Text(title, style: theme.textTheme.labelMedium),
-          ),
-          if (loading)
-            const Positioned(
-              right: 8,
-              top: 8,
-              child: SizedBox(
-                width: 14,
-                height: 14,
-                child: CircularProgressIndicator(strokeWidth: 2),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-
-  IconData _effectIcon(EditorEffectType type) {
-    return switch (type) {
-      EditorEffectType.brightness => Icons.wb_sunny_outlined,
-      EditorEffectType.contrast => Icons.contrast,
-      EditorEffectType.saturation => Icons.palette_outlined,
-      EditorEffectType.temperature => Icons.thermostat,
-      EditorEffectType.gamma => Icons.tune,
-      EditorEffectType.grayscale => Icons.tonality,
-      EditorEffectType.invert => Icons.invert_colors,
-      EditorEffectType.sepia => Icons.filter_vintage,
-      EditorEffectType.denoise => Icons.grain,
-      EditorEffectType.blur => Icons.blur_on,
-      EditorEffectType.sharpen => Icons.auto_fix_high,
-      EditorEffectType.cropToSelection => Icons.crop,
-      EditorEffectType.rotateLeft => Icons.rotate_left,
-      EditorEffectType.rotateRight => Icons.rotate_right,
-      EditorEffectType.flipHorizontal => Icons.swap_horiz,
-      EditorEffectType.flipVertical => Icons.swap_vert,
-    };
-  }
-
-  Future<void> _applyEffect(
-    EditorEffectType effectType,
-    double intensity,
-  ) async {
-    final layer = _state.layerManager.activeLayer;
-    if (layer == null || layer.locked || !layer.hasContent) {
-      AppToast.warning(
-        context,
-        context.l10n.editor_selectUnlockedLayerWithContent,
-      );
-      return;
-    }
-
-    try {
-      final sourceBytes = await _readLayerPng(layer);
-      if (!mounted) return;
-      if (sourceBytes == null) {
-        AppToast.error(context, context.l10n.editor_readCurrentLayerFailed);
-        return;
-      }
-
-      final cropRect = _selectionCropRect();
-      final job = EditorEffectJob(
-        imageBytes: sourceBytes,
-        effectType: effectType,
-        intensity: intensity,
-        cropRect: cropRect,
-      );
-      final resultMessage = await compute(
-        runEditorEffectJobMessage,
-        job.toMessage(),
-        debugLabel: 'image_editor_effect_apply',
-      );
-      final result = EditorEffectResult.fromMessage(resultMessage);
-      final bytes = result.bytes;
-      final newImage = await _decodeUiImage(bytes);
-      if (!mounted) return;
-      _state.historyManager.execute(
-        ReplaceLayerImageAction(
-          layerId: layer.id,
-          newImageBytes: bytes,
-          newImage: newImage,
-          actionDescription: _effectLabel(effectType),
-        ),
-        _state,
-      );
-      _state.layerManager.invalidateSnapshot();
-      setState(() {});
-      AppToast.success(
-        context,
-        context.l10n.editor_effectApplied(_effectLabel(effectType)),
-      );
-    } catch (e) {
-      if (!mounted) return;
-      AppToast.error(context, context.l10n.editor_applyEffectFailed(e));
-    }
-  }
-
-  Future<Uint8List?> _readLayerPng(dynamic layer) async {
-    final rendered = await _renderLayerToImage(layer);
-    try {
-      final raw = await rendered.toByteData(format: ui.ImageByteFormat.png);
-      return raw?.buffer.asUint8List();
-    } finally {
-      rendered.dispose();
-    }
-  }
-
-  Future<ui.Image> _renderLayerToImage(dynamic layer) async {
-    final recorder = ui.PictureRecorder();
-    final canvas = Canvas(recorder);
-    layer.render(canvas, _state.canvasSize);
-    final picture = recorder.endRecording();
-    final image = await picture.toImage(
-      _state.canvasSize.width.toInt(),
-      _state.canvasSize.height.toInt(),
-    );
-    picture.dispose();
-    return image;
-  }
-
-  Future<ui.Image> _decodeUiImage(Uint8List bytes) async {
-    final codec = await ui.instantiateImageCodec(bytes);
-    try {
-      final frame = await codec.getNextFrame();
-      return frame.image;
-    } finally {
-      codec.dispose();
-    }
-  }
-
-  String _effectLabel(EditorEffectType type) {
-    return switch (type) {
-      EditorEffectType.brightness => context.l10n.editor_effectBrightness,
-      EditorEffectType.contrast => context.l10n.editor_effectContrast,
-      EditorEffectType.saturation => context.l10n.editor_effectSaturation,
-      EditorEffectType.temperature => context.l10n.editor_effectTemperature,
-      EditorEffectType.gamma => context.l10n.editor_effectGamma,
-      EditorEffectType.grayscale => context.l10n.editor_effectGrayscale,
-      EditorEffectType.invert => context.l10n.editor_effectInvert,
-      EditorEffectType.sepia => context.l10n.editor_effectSepia,
-      EditorEffectType.denoise => context.l10n.editor_effectDenoise,
-      EditorEffectType.blur => context.l10n.editor_effectBlur,
-      EditorEffectType.sharpen => context.l10n.editor_effectSharpen,
-      EditorEffectType.cropToSelection =>
-        context.l10n.editor_effectCropToSelection,
-      EditorEffectType.rotateLeft => context.l10n.editor_effectRotateLeft,
-      EditorEffectType.rotateRight => context.l10n.editor_effectRotateRight,
-      EditorEffectType.flipHorizontal =>
-        context.l10n.editor_effectFlipHorizontal,
-      EditorEffectType.flipVertical => context.l10n.editor_effectFlipVertical,
-    };
-  }
-
-  double _defaultEffectIntensity(EditorEffectType type) {
-    return editorEffectDefaultIntensity(type);
-  }
-
-  double _effectMin(EditorEffectType type) {
-    return editorEffectMin(type);
-  }
-
-  double _effectMax(EditorEffectType type) {
-    return editorEffectMax(type);
-  }
-
-  bool _effectHasIntensity(EditorEffectType type) {
-    return editorEffectHasIntensity(type);
-  }
-
-  EditorEffectCropRect? _selectionCropRect() {
-    final selection = _state.selectionPath;
-    if (selection == null) {
-      return null;
-    }
-    final bounds = selection.getBounds().intersect(
-      Offset.zero & _state.canvasSize,
-    );
-    if (bounds.isEmpty) {
-      return null;
-    }
-    final x = bounds.left.floor().clamp(0, _state.canvasSize.width - 1).toInt();
-    final y = bounds.top.floor().clamp(0, _state.canvasSize.height - 1).toInt();
-    final right = bounds.right
-        .ceil()
-        .clamp(x + 1, _state.canvasSize.width)
-        .toInt();
-    final bottom = bounds.bottom
-        .ceil()
-        .clamp(y + 1, _state.canvasSize.height)
-        .toInt();
-    return EditorEffectCropRect(
-      x: x,
-      y: y,
-      width: right - x,
-      height: bottom - y,
-    );
-  }
-
-  Widget _buildShortcutSection(String title, List<(String, String)> shortcuts) {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            title,
-            style: theme.textTheme.titleSmall?.copyWith(
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          const SizedBox(height: 8),
-          ...shortcuts.map(
-            (s) => Padding(
-              padding: const EdgeInsets.symmetric(vertical: 2),
-              child: Row(
-                children: [
-                  Container(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 8,
-                      vertical: 2,
-                    ),
-                    decoration: BoxDecoration(
-                      color: theme.colorScheme.surfaceContainerHighest,
-                      borderRadius: BorderRadius.circular(4),
-                    ),
-                    child: Text(
-                      s.$1,
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        fontFamily: 'monospace',
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Text(s.$2, style: theme.textTheme.bodySmall),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// 更改画布尺寸
   Future<void> _changeCanvasSize() async {
     final l10n = context.l10n;
     final result = await CanvasSizeDialog.show(
@@ -2999,291 +1639,6 @@ class _ImageEditorScreenState extends State<ImageEditorScreen> {
     );
   }
 
-  _FocusedInpaintCostEstimate? _resolveFocusedInpaintCostEstimate() {
-    final config = widget.focusedInpaintCostConfig;
-    if (!_isInpaintMode || !_focusedInpaintEnabled || config == null) {
-      return null;
-    }
-
-    final focusAreaRect = _focusedSelectionState.resolveActiveRect(
-      previewPath: _state.previewPath,
-    );
-    if (focusAreaRect == null) {
-      return null;
-    }
-
-    final requestSize = FocusedInpaintUtils.resolveRequestSizeForSelection(
-      sourceWidth: _state.canvasSize.width.round(),
-      sourceHeight: _state.canvasSize.height.round(),
-      selectionRect: focusAreaRect,
-      minContextMegaPixels: _minimumContextMegaPixels,
-    );
-    if (requestSize == null) {
-      return null;
-    }
-
-    final cost = config.estimate(width: requestSize.$1, height: requestSize.$2);
-    if (cost <= 0) {
-      return null;
-    }
-
-    return _FocusedInpaintCostEstimate(
-      width: requestSize.$1,
-      height: requestSize.$2,
-      cost: cost,
-    );
-  }
-
-  Widget _buildFocusedSelectionCard() {
-    final theme = Theme.of(context);
-    final hasFocusArea =
-        _focusedInpaintEnabled && _focusedSelectionState.hasCommittedRect;
-    final costEstimate = _resolveFocusedInpaintCostEstimate();
-
-    return Container(
-      width: 220,
-      padding: const EdgeInsets.all(12),
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surface.withValues(alpha: 0.92),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(
-          color: theme.colorScheme.outline.withValues(alpha: 0.35),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withValues(alpha: 0.16),
-            blurRadius: 16,
-            offset: const Offset(0, 6),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Row(
-            children: [
-              Expanded(
-                child: FilledButton.icon(
-                  onPressed: _toggleFocusedInpaint,
-                  icon: Icon(
-                    _focusedInpaintEnabled
-                        ? Icons.crop_free
-                        : Icons.filter_center_focus,
-                    size: 16,
-                  ),
-                  label: Text(
-                    _focusedInpaintEnabled
-                        ? 'Focused Area Selection'
-                        : 'Focused Inpaint',
-                  ),
-                  style: FilledButton.styleFrom(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 12,
-                      vertical: 10,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Text(
-            !_focusedInpaintEnabled
-                ? context.l10n.editor_focusInactiveHint
-                : hasFocusArea
-                ? context.l10n.editor_focusReadyHint
-                : context.l10n.editor_focusNeedsSelectionHint,
-            style: theme.textTheme.bodySmall,
-          ),
-          if (_focusedInpaintEnabled) ...[
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                _buildFocusModeButton(
-                  icon: Icons.crop_square,
-                  label: context.l10n.editor_focusSelection,
-                  toolId: 'rect_selection',
-                ),
-                const SizedBox(width: 8),
-                _buildFocusModeButton(
-                  icon: Icons.brush_outlined,
-                  label: context.l10n.editor_focusBrush,
-                  toolId: 'brush',
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            Align(
-              alignment: Alignment.centerLeft,
-              child: TextButton.icon(
-                onPressed: _focusedSelectionState.hasCommittedRect
-                    ? () {
-                        setState(() {
-                          _focusedSelectionState.clear();
-                          _state.clearSelection(saveHistory: false);
-                          _state.clearPreview();
-                          _state.setToolById('rect_selection');
-                        });
-                      }
-                    : null,
-                icon: const Icon(Icons.clear, size: 16),
-                label: Text(context.l10n.editor_clearSelection),
-              ),
-            ),
-            const SizedBox(height: 12),
-            Text(
-              context.l10n.editor_focusMinimumContextArea(
-                _minimumContextMegaPixels.round(),
-              ),
-              style: theme.textTheme.labelMedium,
-            ),
-            Slider(
-              value: _minimumContextMegaPixels,
-              min: 0,
-              max: 192,
-              divisions: 192,
-              onChanged: (value) {
-                setState(() {
-                  _minimumContextMegaPixels = value;
-                });
-              },
-            ),
-            if (costEstimate != null) ...[
-              const SizedBox(height: 8),
-              _buildFocusedAnlasWarning(costEstimate),
-              const SizedBox(height: 8),
-            ],
-            Text(
-              context.l10n.editor_focusContextHint,
-              style: theme.textTheme.bodySmall,
-            ),
-          ],
-        ],
-      ),
-    );
-  }
-
-  Widget _buildFocusedAnlasWarning(_FocusedInpaintCostEstimate estimate) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.all(10),
-      decoration: BoxDecoration(
-        color: colorScheme.errorContainer.withValues(alpha: 0.78),
-        borderRadius: BorderRadius.circular(10),
-        border: Border.all(color: colorScheme.error.withValues(alpha: 0.38)),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Icon(
-            Icons.warning_amber_rounded,
-            size: 17,
-            color: colorScheme.onErrorContainer,
-          ),
-          const SizedBox(width: 8),
-          Expanded(
-            child: Text(
-              context.l10n.editor_focusAnlasWarning(
-                estimate.width,
-                estimate.height,
-                estimate.cost,
-              ),
-              style: theme.textTheme.bodySmall?.copyWith(
-                color: colorScheme.onErrorContainer,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _toggleFocusedInpaint() {
-    if (_hasOutpaintChanges && !_focusedInpaintEnabled) {
-      AppToast.warning(
-        context,
-        'Outpaint cannot be used together with Focused Inpaint.',
-      );
-      return;
-    }
-
-    setState(() {
-      _focusedInpaintEnabled = !_focusedInpaintEnabled;
-      if (_focusedInpaintEnabled) {
-        if (!_focusedSelectionState.hasCommittedRect) {
-          _state.setToolById('rect_selection');
-        }
-      } else {
-        _state.clearSelection(saveHistory: false);
-        _state.clearPreview();
-        _focusedSelectionState.clear();
-        _state.setToolById('brush');
-      }
-    });
-  }
-
-  void _consumeFocusedSelection() {
-    if (!_isInpaintMode || !_focusedInpaintEnabled) {
-      return;
-    }
-    if (_state.currentTool?.id != 'rect_selection') {
-      return;
-    }
-    final consumed = _focusedSelectionState.captureSelection(
-      _state.selectionPath,
-    );
-    if (!consumed) {
-      return;
-    }
-
-    _state.clearSelection(saveHistory: false);
-    _state.clearPreview();
-    _state.setToolById('brush');
-    _state.requestUiUpdate();
-    if (mounted) {
-      setState(() {});
-    }
-  }
-
-  Widget _buildFocusModeButton({
-    required IconData icon,
-    required String label,
-    required String toolId,
-  }) {
-    final theme = Theme.of(context);
-    final selected = _state.currentTool?.id == toolId;
-
-    return Expanded(
-      child: OutlinedButton.icon(
-        onPressed: () {
-          _state.setToolById(toolId);
-        },
-        icon: Icon(icon, size: 16),
-        label: Text(label),
-        style: OutlinedButton.styleFrom(
-          foregroundColor: selected
-              ? theme.colorScheme.primary
-              : theme.colorScheme.onSurface,
-          backgroundColor: selected
-              ? theme.colorScheme.primary.withValues(alpha: 0.12)
-              : Colors.transparent,
-          side: BorderSide(
-            color: selected
-                ? theme.colorScheme.primary
-                : theme.colorScheme.outline.withValues(alpha: 0.35),
-          ),
-          padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 10),
-          visualDensity: VisualDensity.compact,
-        ),
-      ),
-    );
-  }
-
   /// 加载蒙版文件
   Future<void> _loadMaskFile() async {
     final l10n = context.l10n;
@@ -3399,51 +1754,5 @@ class _ImageEditorScreenState extends State<ImageEditorScreen> {
   /// 加载蒙版
   Future<void> _loadMask() async {
     await _loadMaskFile();
-  }
-}
-
-class _FocusedContextOverlayPainter extends CustomPainter {
-  _FocusedContextOverlayPainter({
-    required this.canvasController,
-    required this.focusAreaRect,
-    required this.contextCrop,
-    super.repaint,
-  });
-
-  final CanvasController canvasController;
-  final Rect focusAreaRect;
-  final FocusedInpaintCrop contextCrop;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final matrix = canvasController.transformMatrix.storage;
-    final screenSelectionPath = (Path()..addRect(focusAreaRect)).transform(
-      matrix,
-    );
-    final screenContextPath =
-        (Path()..addRect(
-              Rect.fromLTWH(
-                contextCrop.x.toDouble(),
-                contextCrop.y.toDouble(),
-                contextCrop.width.toDouble(),
-                contextCrop.height.toDouble(),
-              ),
-            ))
-            .transform(matrix);
-
-    FocusedOverlayPainter(
-      contextPath: screenContextPath,
-      focusPath: screenSelectionPath,
-    ).paint(canvas, size);
-  }
-
-  @override
-  bool shouldRepaint(covariant _FocusedContextOverlayPainter oldDelegate) {
-    return contextCrop.x != oldDelegate.contextCrop.x ||
-        contextCrop.y != oldDelegate.contextCrop.y ||
-        contextCrop.width != oldDelegate.contextCrop.width ||
-        contextCrop.height != oldDelegate.contextCrop.height ||
-        focusAreaRect != oldDelegate.focusAreaRect ||
-        canvasController != oldDelegate.canvasController;
   }
 }
